% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qapreg.R
\name{coef.qapreg}
\alias{coef.qapreg}
\title{Extract QAP regression coefficients}
\usage{
\method{coef}{qapreg}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"qapreg"}.}

\item{...}{Additional arguments, currently ignored.}
}
\value{
A named numeric vector of coefficients.
}
\description{
Extracts observed model coefficients from a fitted \code{qapreg} object.
}
