% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qapreg.R
\name{.qap_extract_network_attributes}
\alias{.qap_extract_network_attributes}
\title{Add vertex and edge attributes to dyadic data}
\usage{
.qap_extract_network_attributes(net, dyads, dat)
}
\arguments{
\item{net}{A \pkg{network} object.}

\item{dyads}{A data frame with integer columns \code{i} and \code{j} identifying dyads.}

\item{dat}{A dyadic data frame to which attributes will be added.}
}
\value{
The input data frame with additional attribute columns.
}
\description{
Extracts vertex attributes as sender/receiver variables and edge attributes
as dyadic variables.
}
\keyword{internal}
