% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qapreg.R
\name{.qap_fit}
\alias{.qap_fit}
\title{Fit a model using a supplied fitting function}
\usage{
.qap_fit(formula, data, fit_fun, fit_args)
}
\arguments{
\item{formula}{A model formula.}

\item{data}{A data frame used for model estimation.}

\item{fit_fun}{A model-fitting function such as \code{\link[stats:glm]{stats::glm()}} or
\code{\link[stats:lm]{stats::lm()}}.}

\item{fit_args}{A named list of additional arguments passed to \code{fit_fun}.}
}
\value{
A fitted model object.
}
\description{
Calls a model-fitting function with a formula, data frame, and additional
arguments.
}
\keyword{internal}
