% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qapreg.R
\name{qapreg-package}
\alias{qapreg-package}
\title{qapreg: QAP regression for statnet network objects}
\description{
Tools for preparing dyadic data from \pkg{network} objects and estimating
regression models with Quadratic Assignment Procedure (QAP) inference.
}
\keyword{internal}
